work_with_storage/python/requirements.txt
azure-core==1.29.4
azure-identity==1.14.0
azure-storage-blob==12.18.1
azure-storage-queue==12.7.1
isodate==0.6.1
msal==1.24.0
msal-extensions==1.0.0
portalocker==2.7.0
PyJWT==2.8.0
